added MetaToolTip component; added MetaToolTip to dynamic and static mfd dialogs;
